Add synthetic vectors support for rank_vectors
This change adds the support for synthetic vectors (added in #130382) in the rank_vectors field type.
