+++ x-pack/plugin/rank-vectors/src/main/java/org/elasticsearch/xpack/rank/vectors/RankVectorsPlugin.java
@@ -20,6 +20,7 @@ import org.elasticsearch.xpack.rank.vectors.mapper.RankVectorsFieldMapper;
 
 import java.util.Map;
 
+import static org.elasticsearch.index.IndexSettings.INDEX_MAPPING_SOURCE_SYNTHETIC_VECTORS_SETTING;
 import static org.elasticsearch.index.mapper.FieldMapper.notInMultiFields;
 import static org.elasticsearch.xpack.rank.vectors.mapper.RankVectorsFieldMapper.CONTENT_TYPE;
 
@@ -36,7 +37,12 @@ public class RankVectorsPlugin extends Plugin implements MapperPlugin {
             if (RANK_VECTORS_FEATURE.check(getLicenseState()) == false) {
                 throw LicenseUtils.newComplianceException("Rank Vectors");
             }
-            return new RankVectorsFieldMapper.Builder(n, c.indexVersionCreated(), getLicenseState());
+            return new RankVectorsFieldMapper.Builder(
+                n,
+                c.indexVersionCreated(),
+                getLicenseState(),
+                INDEX_MAPPING_SOURCE_SYNTHETIC_VECTORS_SETTING.get(c.getIndexSettings().getSettings())
+            );
         }, notInMultiFields(CONTENT_TYPE)));
     }
